# Backend/Frontend/openapi.raw.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/tickets": {
      "post": {
        "operationId": "TicketsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Insufficient role permissions"
          },
          "404": {
            "description": "Category not found (or inactive)"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a new ticket",
        "tags": [
          "tickets"
        ]
      },
      "get": {
        "operationId": "TicketsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TicketStatus"
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/TicketPriority"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "assigneeId",
            "required": false,
            "in": "query",
            "description": "Honored for ADMIN callers only",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "createdById",
            "required": false,
            "in": "query",
            "description": "Honored for ADMIN callers only",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text search on title and reference (case-insensitive)",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "priority",
                "slaDueAt",
                "status"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TicketListItemDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List tickets, paginated and filtered, scoped by the caller's role",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/tickets/{id}": {
      "get": {
        "operationId": "TicketsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not the owner, the assignee, nor an admin"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a single ticket by id",
        "tags": [
          "tickets"
        ]
      },
      "patch": {
        "operationId": "TicketsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "No field provided to update"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not an admin, not the owning CLIENT with the ticket still OPEN (nor the owner at all), or a TECHNICIAN"
          },
          "404": {
            "description": "Ticket not found, or the given categoryId is unknown/inactive"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update mutable fields of a ticket (role x status rules enforced by the service)",
        "tags": [
          "tickets"
        ]
      },
      "delete": {
        "operationId": "TicketsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Ticket soft deleted"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Caller is not an ADMIN\n\nInsufficient role permissions"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Soft delete a ticket (ADMIN only)",
        "tags": [
          "tickets"
        ]
      }
    },
...
    "/api/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, or an unknown field was sent"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Caller is not an ADMIN\n\nInsufficient role permissions"
          },
          "404": {
            "description": "The given requiredSkillId is unknown"
          },
          "409": {
            "description": "A category with this name already exists"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a ticket category (ADMIN only)",
        "tags": [
          "categories"
        ]
      },
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by Category.isActive. A ticket-creation form wants isActive=true.",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List categories, sorted by name ASC (not paginated). Available to every authenticated role.",
        "tags": [
          "categories"
        ]
      }
    },
    "/api/categories/{id}": {
      "get": {
        "operationId": "CategoriesController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a single category by id",
        "tags": [
          "categories"
        ]
      },
      "patch": {
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "No field provided to update"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Caller is not an ADMIN\n\nInsufficient role permissions"
          },
          "404": {
            "description": "Category not found, or the given requiredSkillId is unknown"
          },
          "409": {
            "description": "Another category already uses this name"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a category — including retiring it with isActive: false, which is how a category is removed (there is no DELETE, D6)",
        "tags": [
          "categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ticket Checker API",
    "description": "",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 42,
            "description": "Total number of items across all pages"
          },
          "page": {
            "type": "number",
            "example": 1,
            "description": "Current, 1-based page number"
          },
          "limit": {
            "type": "number",
            "example": 20,
            "description": "Number of items per page"
          },
          "totalPages": {
            "type": "number",
            "example": 3,
            "description": "Total number of pages"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "PaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Page of results. The concrete item type is documented per-endpoint by @ApiPaginatedResponse()."
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "TicketPriority": {
        "type": "string",
        "enum": [
          "LOW",
          "NORMAL",
          "HIGH",
          "CRITICAL"
        ]
      },
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Panne de climatisation salle serveur",
            "minLength": 3,
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "example": "La climatisation ne démarre plus depuis ce matin.",
            "minLength": 1,
            "maxLength": 5000
          },
          "priority": {
            "default": "NORMAL",
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketPriority"
              }
            ]
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "siteLabel": {
            "type": "string",
            "maxLength": 150
          },
          "siteAddress": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "title",
          "description",
          "categoryId"
        ]
      },
      "TicketStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "ASSIGNED",
          "IN_PROGRESS",
          "RESOLVED",
          "CLOSED",
          "CANCELLED"
        ]
      },
      "TicketCategorySummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "UserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "firstName",
          "lastName"
        ]
      },
      "TicketResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketStatus"
              }
            ]
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketPriority"
              }
            ]
          },
          "category": {
            "$ref": "#/components/schemas/TicketCategorySummaryDto"
          },
          "createdBy": {
            "$ref": "#/components/schemas/UserSummaryDto"
          },
          "assignee": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserSummaryDto"
              }
            ]
          },
          "siteLabel": {
            "type": "string",
            "nullable": true
          },
          "siteAddress": {
            "type": "string",
            "nullable": true
          },
          "slaDueAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "assignedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "startedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "resolvedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "closedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "cancelledAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "resolutionNote": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reference",
          "title",
          "description",
          "status",
          "priority",
          "category",
          "createdBy",
          "assignee",
          "siteLabel",
          "siteAddress",
          "slaDueAt",
          "assignedAt",
          "startedAt",
          "resolvedAt",
          "closedAt",
          "cancelledAt",
          "resolutionNote",
          "createdAt",
          "updatedAt"
        ]
      },
      "TicketListCategorySummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "TicketListAssigneeSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
      "TicketListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketStatus"
              }
            ]
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketPriority"
              }
            ]
          },
          "category": {
            "$ref": "#/components/schemas/TicketListCategorySummaryDto"
          },
          "assignee": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketListAssigneeSummaryDto"
              }
            ]
          },
          "slaDueAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reference",
          "title",
          "status",
          "priority",
          "category",
          "assignee",
          "slaDueAt",
          "createdAt"
        ]
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Panne de climatisation salle serveur",
            "minLength": 3,
            "maxLength": 150
          },
          "description": {
            "type": "string",
            "example": "La climatisation ne démarre plus depuis ce matin.",
            "minLength": 1,
            "maxLength": 5000
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TicketPriority"
              }
            ]
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "siteLabel": {
            "type": "string",
            "maxLength": 150
          },
          "siteAddress": {
            "type": "string",
            "maxLength": 2000
          }
        }
      },
...
      "SkillResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ]
      },
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Panne électrique",
            "minLength": 2,
            "maxLength": 80
          },
          "description": {
            "type": "string",
            "example": "Coupure ou dysfonctionnement électrique.",
            "maxLength": 2000
          },
          "requiredSkillId": {
            "type": "string",
            "format": "uuid",
            "description": "Skill required to handle this category. Drives the automatic assignment suggestion."
          }
        },
        "required": [
          "name"
        ]
      },
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "requiredSkill": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/SkillResponseDto"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "requiredSkill",
          "isActive"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "requiredSkillId": {
            "type": "object",
            "format": "uuid",
            "nullable": true,
            "description": "Send null to clear the required skill; omit to leave it as is."
          },
          "isActive": {
            "type": "boolean",
            "description": "Set to false to retire the category: it stays attached to historical tickets but can no longer be chosen for a new one."
          }
        }
      },
...
    }
  }
}